Revert "chore: give better names to logo refs and divs"
This reverts commit 6edfa8ad7446c6ea030bc1d1f7dbb662859f8ba0.
