+++ client/src/assets/logos/LogoTextSide.tsx
@@ -8,7 +8,7 @@ interface ILogoProps extends SVGAttributes<Element> {
 
 const LogoTextSide: React.FunctionComponent<ILogoProps> = ({ fill = '#000', size = 4, style, ...props }) => {
   style = {
-    height: size + 'rem',
+    height: size - 4 + 'rem',
     width: size + 'rem',
     ...style,
   }
@@ -17,7 +17,7 @@ const LogoTextSide: React.FunctionComponent<ILogoProps> = ({ fill = '#000', size
     <svg
       width={style.width}
       height={style.height}
-      viewBox="0 0 66 92"
+      viewBox="0 0 160 61"
       xmlns="http://www.w3.org/2000/svg"
       style={style}
       {...props}
@@ -25,7 +25,7 @@ const LogoTextSide: React.FunctionComponent<ILogoProps> = ({ fill = '#000', size
       {' '}
       <path
         fill="#fff"
-        d="M35.912 7.702c13.818 0 25.02 11.237 25.02 25.098 0 13.862-11.202 25.098-25.02 25.098S10.892 46.662 10.892 32.8c0-.78.035-1.551.104-2.313.342.021.686.032 1.033.032 9.421 0 17.06-7.662 17.06-17.113 0-1.585-.216-3.12-.618-4.576 2.35-.733 4.85-1.128 7.44-1.128zm3.235 12.55l-4.408.019-7.925 21.424 3.443 3.029 11.409-5.364 1.87 5.134 3.905-1.416-8.294-22.827zm-4.157.697l.232.718-7.22 19.53-.589.273 7.577-20.52zm.483 1.397l.232.698-6.524 17.609-.575.266 6.867-18.573zm.464 1.396l.232.679-5.815 15.69-.605.279 6.188-16.648zm.484 1.377l.232.698-5.123 13.75-.58.268 5.47-14.716zm.527 1.287l3.289 9.03-8.106 3.852 4.817-12.882zM12.823 8.164v18.29H8.724V8.164h4.099zM7.917 6.692v18.29H3.818V6.692h4.099zm9.677-.848v18.29h-4.098V5.844h4.098zm4.772-3.236v18.29h-4.1V2.607h4.1zM2.177 91.671l.4-.867h3.265l.413.867h2.177L4.382 83h-.359L0 91.671h2.177zm2.976-2.335H3.21l.965-2.126h.027l.95 2.126zm6.383 2.335v-4.993l3.224 4.993h2.108V83.28h-2.012v4.979l-3.21-4.98H9.511v8.392h2.025zm8.588 0l.4-.867h3.265l.413.867h2.177L22.328 83h-.358l-4.023 8.671h2.177zm2.976-2.335h-1.943l.964-2.126h.028l.95 2.126zm9.979 2.335V89.91h-3.596v-6.63h-2.026v8.392h5.622zm4.04 0v-4.056l2.935-4.335h-2.273l-1.667 2.517-1.654-2.517h-2.287l2.935 4.335v4.056h2.012zm7.61 0v-6.573h1.874V83.28h-5.76v1.818h1.847v6.573h2.039zm5.033 0V83.28h-2.026v8.391h2.026zm6.08.28c1.53 0 2.673-.658 3.5-1.539l-1.35-1.342c-.58.573-1.324.937-2.15.937-1.474 0-2.425-1.119-2.425-2.532 0-1.412.95-2.531 2.425-2.531.826 0 1.57.364 2.15.937l1.39-1.343c-.881-.937-1.97-1.538-3.54-1.538-2.618 0-4.478 1.972-4.478 4.475 0 2.504 1.86 4.476 4.478 4.476zm7.361 0c1.557 0 2.797-1.23 2.797-2.713 0-1.315-.689-2.154-2.163-2.602-.703-.21-1.46-.321-1.46-.95 0-.546.427-.812.867-.812.565 0 .855.392.855.867H66C65.89 84.19 65.022 83 63.19 83c-1.558 0-2.811 1.23-2.811 2.713 0 1.44.771 2.098 2.163 2.602.675.237 1.488.42 1.488 1.048 0 .434-.317.728-.827.728-.468 0-.854-.392-.854-.881h-1.97c0 1.51 1.02 2.74 2.824 2.74zM7.019 78.671v-3.133l2.1 3.133h2.528l-2.225-3.119c1.222-.307 1.95-1.133 1.95-2.601 0-1.58-1.154-2.671-2.925-2.671H5v8.391h2.019zm1.62-4.755H7.02v-1.888h1.565c.453 0 .81.433.81.923 0 .49-.302.965-.755.965zm9.521 4.755V76.91h-3.584v-1.51h3.172v-1.79h-3.172v-1.511h3.433V70.28h-5.452v8.391h5.603zm2.902 0l.398-.867h3.254l.412.867h2.17L23.26 70h-.357l-4.01 8.671h2.17zm2.966-2.335H22.09l.962-2.126h.027l.948 2.126zm7.653 2.615c1.524 0 2.664-.658 3.488-1.539l-1.346-1.342c-.577.573-1.318.937-2.142.937-1.47 0-2.417-1.119-2.417-2.532 0-1.412.948-2.531 2.417-2.531.824 0 1.565.364 2.142.937l1.387-1.343C34.331 70.601 33.246 70 31.68 70c-2.608 0-4.462 1.972-4.462 4.475 0 2.504 1.854 4.476 4.463 4.476zm8.189-.28v-6.573h1.867V70.28h-5.74v1.818h1.84v6.573h2.033zm5.016 0V70.28h-2.019v8.391h2.019zm4.88.28l3.638-8.671h-2.1l-1.662 4.042-1.662-4.042h-2.101l3.639 8.67h.247zM60 78.67V76.91h-3.584v-1.51h3.172v-1.79h-3.172v-1.511h3.433V70.28h-5.452v8.391H60z"
+        d="M32.8 7.45C46.165 7.45 57 18.318 57 31.725 57 45.132 46.165 56 32.8 56 19.435 56 8.6 45.132 8.6 31.725c0-.754.034-1.5.101-2.237.331.02.664.03.999.03 9.113 0 16.5-7.41 16.5-16.551 0-1.533-.208-3.017-.596-4.426 2.273-.71 4.69-1.091 7.196-1.091zm3.129 12.137l-4.264.02L24 40.326l3.33 2.93 11.035-5.188 1.809 4.966 3.777-1.37-8.022-22.078zm-4.02.675l.224.695-6.984 18.89-.569.263 7.328-19.848zm.467 1.351l.224.675-6.31 17.032-.557.257 6.643-17.964zm.449 1.35l.224.657-5.624 15.175-.586.27 5.986-16.101zm.467 1.333l.224.675-4.954 13.298-.562.26 5.292-14.233zm.51 1.244l3.181 8.733L29.143 38l4.66-12.46zM9.477 7.661v15.242H6.06V7.66h3.416zM5.387 6.434v15.242H1.972V6.434h3.415zm8.065-.706V20.97h-3.416V5.728h3.416zm3.976-2.697v15.241h-3.416V3.031h3.416zM62.166 46.594l.582-1.26h4.75l.6 1.26h3.167L65.373 34h-.521L59 46.594h3.166zm4.33-3.392h-2.827l1.403-3.088h.04l1.383 3.088zm9.284 3.392v-7.252l4.69 7.252h3.065V34.406H80.61v7.232l-4.669-7.232h-3.106v12.188h2.946zm12.49 0l.582-1.26h4.75l.6 1.26h3.167L91.477 34h-.52l-5.853 12.594h3.167zm4.33-3.392h-2.826l1.403-3.088h.04l1.382 3.088zm14.515 3.392v-2.56h-5.231v-9.628h-2.946v12.188h8.177zm5.877 0v-5.89l4.269-6.298h-3.307l-2.425 3.657-2.405-3.657h-3.326l4.268 6.297v5.89h2.926zm11.068 0v-9.547h2.726v-2.64h-8.377v2.64h2.685v9.547h2.966zm7.32 0V34.406h-2.945v12.188h2.946zm8.844.406c2.225 0 3.888-.955 5.09-2.234l-1.963-1.95c-.842.832-1.924 1.36-3.127 1.36-2.144 0-3.527-1.624-3.527-3.676s1.383-3.677 3.527-3.677c1.203 0 2.285.529 3.127 1.361l2.024-1.95c-1.283-1.36-2.866-2.234-5.15-2.234-3.808 0-6.514 2.864-6.514 6.5s2.706 6.5 6.513 6.5zm10.708 0c2.264 0 4.068-1.788 4.068-3.94 0-1.91-1.002-3.129-3.146-3.779-1.022-.304-2.125-.467-2.125-1.381 0-.792.622-1.178 1.263-1.178.822 0 1.242.569 1.242 1.26H155c-.16-2.255-1.423-3.982-4.088-3.982-2.265 0-4.089 1.788-4.089 3.94 0 2.093 1.123 3.047 3.147 3.779.982.345 2.164.61 2.164 1.523 0 .63-.46 1.056-1.202 1.056-.682 0-1.243-.568-1.243-1.28h-2.866c0 2.194 1.483 3.982 4.109 3.982zM66.936 27.594v-4.55l3.057 4.55h3.675l-3.236-4.53c1.778-.447 2.836-1.645 2.836-3.778 0-2.295-1.677-3.88-4.254-3.88H64v12.188h2.936zm2.357-6.907h-2.357v-2.742h2.277c.66 0 1.179.63 1.179 1.34 0 .712-.44 1.402-1.099 1.402zm13.849 6.907v-2.56h-5.214v-2.193h4.615v-2.6h-4.615v-2.194h4.994v-2.64h-7.93v12.187h8.15zm4.22 0l.58-1.26h4.734l.599 1.26h3.156L90.558 15h-.52l-5.832 12.594h3.156zm4.315-3.392H88.86l1.399-3.088h.04l1.378 3.088zM102.809 28c2.217 0 3.875-.955 5.073-2.234l-1.957-1.95c-.84.832-1.918 1.36-3.116 1.36-2.138 0-3.516-1.624-3.516-3.676s1.378-3.677 3.516-3.677c1.198 0 2.277.529 3.116 1.361l2.017-1.95c-1.278-1.36-2.856-2.234-5.133-2.234-3.796 0-6.492 2.864-6.492 6.5s2.696 6.5 6.492 6.5zm11.91-.406v-9.547h2.717v-2.64h-8.35v2.64h2.677v9.547h2.956zm7.297 0V15.406h-2.936v12.188h2.936zm7.097.406l5.293-12.594h-3.056l-2.417 5.87-2.417-5.87h-3.056L128.753 28h.36zM144 27.594v-2.56h-5.213v-2.193h4.614v-2.6h-4.614v-2.194h4.993v-2.64h-7.93v12.187H144z"
       />
     </svg>
   )
