--- client/src/assets/logos/LogoTextTop.tsx
@@ -1,38 +1,0 @@
-import React, { SVGAttributes } from 'react'
-import styled from 'styled-components/macro'
-
-interface ILogoProps extends SVGAttributes<Element> {
-  fill?: string
-  size?: number
-}
-
-const LogoTextTop: React.FunctionComponent<ILogoProps> = ({ fill = '#000', size = 4, style, ...props }) => {
-  style = {
-    height: size - 4 + 'rem',
-    width: size + 'rem',
-    ...style,
-  }
-
-  return (
-    <svg
-      width={style.width}
-      height={style.height}
-      viewBox="0 0 160 61"
-      xmlns="http://www.w3.org/2000/svg"
-      style={style}
-      {...props}
-    >
-      {' '}
-      <path
-        fill="#fff"
-        d="M32.8 7.45C46.165 7.45 57 18.318 57 31.725 57 45.132 46.165 56 32.8 56 19.435 56 8.6 45.132 8.6 31.725c0-.754.034-1.5.101-2.237.331.02.664.03.999.03 9.113 0 16.5-7.41 16.5-16.551 0-1.533-.208-3.017-.596-4.426 2.273-.71 4.69-1.091 7.196-1.091zm3.129 12.137l-4.264.02L24 40.326l3.33 2.93 11.035-5.188 1.809 4.966 3.777-1.37-8.022-22.078zm-4.02.675l.224.695-6.984 18.89-.569.263 7.328-19.848zm.467 1.351l.224.675-6.31 17.032-.557.257 6.643-17.964zm.449 1.35l.224.657-5.624 15.175-.586.27 5.986-16.101zm.467 1.333l.224.675-4.954 13.298-.562.26 5.292-14.233zm.51 1.244l3.181 8.733L29.143 38l4.66-12.46zM9.477 7.661v15.242H6.06V7.66h3.416zM5.387 6.434v15.242H1.972V6.434h3.415zm8.065-.706V20.97h-3.416V5.728h3.416zm3.976-2.697v15.241h-3.416V3.031h3.416zM62.166 46.594l.582-1.26h4.75l.6 1.26h3.167L65.373 34h-.521L59 46.594h3.166zm4.33-3.392h-2.827l1.403-3.088h.04l1.383 3.088zm9.284 3.392v-7.252l4.69 7.252h3.065V34.406H80.61v7.232l-4.669-7.232h-3.106v12.188h2.946zm12.49 0l.582-1.26h4.75l.6 1.26h3.167L91.477 34h-.52l-5.853 12.594h3.167zm4.33-3.392h-2.826l1.403-3.088h.04l1.382 3.088zm14.515 3.392v-2.56h-5.231v-9.628h-2.946v12.188h8.177zm5.877 0v-5.89l4.269-6.298h-3.307l-2.425 3.657-2.405-3.657h-3.326l4.268 6.297v5.89h2.926zm11.068 0v-9.547h2.726v-2.64h-8.377v2.64h2.685v9.547h2.966zm7.32 0V34.406h-2.945v12.188h2.946zm8.844.406c2.225 0 3.888-.955 5.09-2.234l-1.963-1.95c-.842.832-1.924 1.36-3.127 1.36-2.144 0-3.527-1.624-3.527-3.676s1.383-3.677 3.527-3.677c1.203 0 2.285.529 3.127 1.361l2.024-1.95c-1.283-1.36-2.866-2.234-5.15-2.234-3.808 0-6.514 2.864-6.514 6.5s2.706 6.5 6.513 6.5zm10.708 0c2.264 0 4.068-1.788 4.068-3.94 0-1.91-1.002-3.129-3.146-3.779-1.022-.304-2.125-.467-2.125-1.381 0-.792.622-1.178 1.263-1.178.822 0 1.242.569 1.242 1.26H155c-.16-2.255-1.423-3.982-4.088-3.982-2.265 0-4.089 1.788-4.089 3.94 0 2.093 1.123 3.047 3.147 3.779.982.345 2.164.61 2.164 1.523 0 .63-.46 1.056-1.202 1.056-.682 0-1.243-.568-1.243-1.28h-2.866c0 2.194 1.483 3.982 4.109 3.982zM66.936 27.594v-4.55l3.057 4.55h3.675l-3.236-4.53c1.778-.447 2.836-1.645 2.836-3.778 0-2.295-1.677-3.88-4.254-3.88H64v12.188h2.936zm2.357-6.907h-2.357v-2.742h2.277c.66 0 1.179.63 1.179 1.34 0 .712-.44 1.402-1.099 1.402zm13.849 6.907v-2.56h-5.214v-2.193h4.615v-2.6h-4.615v-2.194h4.994v-2.64h-7.93v12.187h8.15zm4.22 0l.58-1.26h4.734l.599 1.26h3.156L90.558 15h-.52l-5.832 12.594h3.156zm4.315-3.392H88.86l1.399-3.088h.04l1.378 3.088zM102.809 28c2.217 0 3.875-.955 5.073-2.234l-1.957-1.95c-.84.832-1.918 1.36-3.116 1.36-2.138 0-3.516-1.624-3.516-3.676s1.378-3.677 3.516-3.677c1.198 0 2.277.529 3.116 1.361l2.017-1.95c-1.278-1.36-2.856-2.234-5.133-2.234-3.796 0-6.492 2.864-6.492 6.5s2.696 6.5 6.492 6.5zm11.91-.406v-9.547h2.717v-2.64h-8.35v2.64h2.677v9.547h2.956zm7.297 0V15.406h-2.936v12.188h2.936zm7.097.406l5.293-12.594h-3.056l-2.417 5.87-2.417-5.87h-3.056L128.753 28h.36zM144 27.594v-2.56h-5.213v-2.193h4.614v-2.6h-4.614v-2.194h4.993v-2.64h-7.93v12.187H144z"
-      />
-    </svg>
-  )
-}
-
-export default styled(LogoTextTop)`
-  [fill] {
-    fill: ${({ theme }) => theme.primary.corePrimary};
-  }
-`
