+++ client/src/assets/logos/index.ts
@@ -1,2 +1,2 @@
+export { default as LogoTextBottom } from './LogoTextBottom'
 export { default as LogoTextSide } from './LogoTextSide'
-export { default as LogoTextTop } from './LogoTextTop'
